notice 오류 수정
